backend/pyproject.toml
[project]
name = "mogged-tv-backend"
version = "0.1.0"
description = "Backend API for mogged.tv — a private live streaming platform"
authors = [
    { name = "srgeddes", email = "rileygeddes22@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "pydantic-settings",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "alembic",
    "greenlet",
    "bcrypt",
    "pyjwt",
    "livekit-api>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.sources]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.format]
indent-style = "space"
quote-style = "double"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "ASYNC", "W", "C4", "PIE", "RET", "RSE"]
ignore = ["UP046"]

[tool.ruff.lint.isort]
known-first-party = ["core", "auth", "streams", "users", "livekit", "organizations", "friends", "trivia"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[dependency-groups]
dev = [
    "ruff>=0.15.4",
    "pytest",
    "pytest-asyncio",
    "httpx>=0.28.1",
    "aiosqlite>=0.22.1",
]
